Test CalculatePercentage

+++ CreditModule.DomainModels/Payment.cs
@@ -4,7 +4,7 @@ namespace CreditModule.DomainModels;
 
 public class Payment
 {
-    public int Sum { get; set; }
+    public decimal Sum { get; set; }
     public PaymentTypes PaymentType { get; set; }
 
 }
